refactor!: Change sample format to always include PID
The format of the samples emitted by Austin has been modified to always 
show the PID using the shorter format P<pid>. The thread label has been 
shortened as well to T<tid>.

BREAKING CHANGE: consumers must be aware of the new format

+++ src/py_thread.h
@@ -23,6 +23,7 @@
 #ifndef PY_THREAD_H
 #define PY_THREAD_H
 
+#include <stdint.h>
 #include <sys/types.h>
 
 #include "mem.h"
@@ -35,7 +36,7 @@ typedef struct thread {
   raddr_t         raddr;
   raddr_t         next_raddr;
 
-  long            tid;
+  uintptr_t       tid;
   struct thread * next;
 
   py_frame_t    * first_frame;
